submit

+++ libft/ft_bzero.c
@@ -1,3 +1,5 @@
+#include<unistd.h>
+
 void	ft_bzero(void *s, size_t n)
 {
 	int		i;
@@ -7,9 +9,27 @@ void	ft_bzero(void *s, size_t n)
 	i = 0;
 	if (n == 0)
 		return ;
-	while (i < n)
+	while (i < (int)n)
 	{
 		str[i] = 0;
 		i++;
 	}
 }
+
+// #include<string.h>
+// #include<stdio.h>
+// int main() {
+//     char buffer[20];
+
+//     // テストケース 1: バッファをゼロで埋める
+//     strcpy(buffer, "Hello, World!");
+//     ft_bzero(buffer, sizeof(buffer));
+//     printf("Test Case 1: %s\n", buffer);  // 期待される出力: 空文字列
+
+//     // テストケース 2: バッファの一部をゼロで埋める
+//     strcpy(buffer, "Hello, World!");
+//     ft_bzero(buffer, 2);
+//     printf("Test Case 2: %s\n", buffer);  // 期待される出力: "Hello,\0\0\0\0\0orld!"
+
+//     return 0;
+// }
